BackendWithFronted folder added.

+++ Backend/Interfaces/IGenericService.cs
@@ -1,10 +1,11 @@
 using Backend.Models;
 namespace Backend.Interfaces;
+
 public interface IGenericService<T>
 {
     Task<Response<IEnumerable<T>>> GetAllAsync();
-    Task<Response<T>> GetByIdAsync(int id);
+    Task<Response<T>> GetByIdAsync(long id);
     Task<Response<T>> AddAsync(T entity);
-    Task<Response<T>> UpdateAsync(int id, T entity);
-    Task<Response<bool>> DeleteAsync(int id);
+    Task<Response<T>> UpdateAsync(long id, T entity);
+    Task<Response<bool>> DeleteAsync(long id);
 }
